pyproject.toml
[tool.poetry]
name = "MLOps-Project"
version = "0.1.0"
description = "MLOps project"
authors = ["Palandr123 <palandr545@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
dvc = "^3.51.2"
dvc-gdrive = "^3.0.1"
pandas = "^2.2.2"
numpy = "<2.0.0"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
scikit-learn = "^1.5.0"
category-encoders = "^2.6.3"
hydra-core = "^1.3.2"
kaggle = "^1.6.14"
pytest = "^8.2.2"
great-expectations = "^0.18.16"
zenml = {extras = ["server"], version = "0.50.0"}
apache-airflow = { extras = ["cncf.kubernetes"], version = "2.7.3" }
flask-session2 = "^1.3.1"
psycopg2-binary = "^2.9.9"
mlflow = "^2.14.2"
yq = "^3.4.3"
torch = {version = "2.1.2+cu121", source = "pytorch-cu121-src"}
skorch = "^1.0.0"


jq = "^1.7.0"
giskard = "2.14.0"
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cu121-src"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"
